I wrote a new function: read_data2
read_data2 is using all the information given by excell to provide the final data_set. In particular, dataset_main (now slightly modified) provides names, trnasformations, in/out, and number of the series. Using this info it automatically truncates the dataset and remove all the variables which are not part of the VAR using the information in_or_out. In in_or_out = 1 the variable is inside the VAR if In_or_out is equal to zero the variable is out. This function does NOT provide which_shocks and position_rel_prices. In my opinion this info is too exercise specific and I guess we may formally add them in the main file. I have compared the old data with the new data and the result is the same, but we still nedd to double check if the result is correct.