Dataset_test_PC almost completed
Now we have 118 variables. We basically miss 5 variables which I had to apply for a dataset to have them and I am waiting for the confirmation, Shiller variables that it is not clear if they are important (I will dig into that) and that's it. Moreover, the formula for the relative of the michigan index is simply: Better - Worse + 100.