How to transform Data
I found a function of forni created the same day when the Dataset was modified that read and transform data. The dataset was modified last time that day so my interpretation is that the function and the dataset speak the same language. We can map Forni's transformation for our dataset and we are ready to take the test