Forni&Gambetti Test Implemented
As a first look results are strongly encouraging.

IT shock super easily pass the test. 99.69%

News shock reject a little bit 77.18%

However, this is a super good news in my opinion.  Forni&Gambetti invented the test for specific shocks, such as news shock, that needs A LOT of information to be pin down. The fact that IT shock is easily passing the test and, on the other hand, the news shock is not, suggests that our identification strategy is correct. Indeed is the news shock the one with information trouble and not the It shock!