correct_favourite_macrolunch
A lot of things are going on.

1. I have updated the figures in order to allow 2 confidence intervals (90% and 95% for example)

2. Very inelegantly I have fixed the scale of the IRFs across shocks. You will see how I have done looking at plot_single_IRF

3. I run 1000 bootstraps of the new corrected specification (Real SP500 are now in logs), 1 lag, 8 is the horizon contraint.

Problems.

For some obscure reasons I do not know why the point estimate is not anymore perfectly inside the confidence intervals. You can see such results on figures that I run todat at 2pm (precisely).

Which are the problems?

1. Code mistake.

Solutions.
a. Comparing the old code with this one.
b.Re-run the uncorrected specification.

2. Nonstationarity.

Solutions.
a. Need to check at very long horizon if the bootstrap is exploding.
b. Using a slightly different specification.