No progress
I used dbstop and apparently the problem appears since the beginning. It
is not a problem of plotting or rewriting. I tool gam3_opt just
evaluated and then I build up IRFs and plot it with a simple
plot(ir(2,:,6)). The problem is either in the constraint, somehow has
fucked up or in genIRFs. I cannot think the problem is somewhere else.