Just downloaded in a separate file and then readded in correct_favourite_macrolunch.mat
I had the problematic 1000 bootstrap workspace. No w it is in and you
can play around. I checked how was working using the old code (to
genIRFs and to plot it) but it seems that the confidence interval
problem is still there. It is not orrible but still not nicely in the
middle. As it is in the tex presentation.