a workspace with log(cumsum(TFP)) when we read in data.
Instead of cumsum(TFP). Why? Because I noticed that our TFP series is
massively bigger than all the other series. With this extra log, the
TFP IRFs are the same magnitude as the responses of the other shocks.
The 2nd good news is that this magnitude is the same as that of Barsky
and Sims.