Plotting IRFs
The final beta matrix to obtain the IRF is non stationary. I need to understand if it is a feature of the final form of the VECM of I have done something wrong. Short-run IRFs are smooth and nice...

In any case we need to speak with Ryan on Wednesday