Improvement over VECM
I think I understood the problem if fmincon. The problem is our restrictions. We cannot hope to always obtain a salution if the number of cointegrating vectors is not correct. Data must agree on our imposition otherwise the exercise as either no solution or corner solution. Maybe there is some hope to make it run before Ryan.