a way to calc SS without eq 24 from SS pdf
Eq 24 is where 1/gam comes in. Clearly, when gam=0, this makes the
objective function undefined.
Using this new code we get an unfortunately imprecise solution for the
steady state, but at least the model works for gam=0 as well. I’ll redo
the IR-matching in this case to see what value of gamma we get.